Refactor code based on previous comments
