+++ AirportTicketBookingSystem/FlightManagement/Booking.cs
@@ -33,12 +33,6 @@ namespace AirportTicketBookingSystem.FlightManagement
             return stringBuilder.ToString();
         }
 
-        public static string SaveToFile(Booking b)
-        {
-            StringBuilder stringBuilder = new StringBuilder();
-            stringBuilder.AppendLine($"{b.Id};{b.Flight.Id};{(int)b.FlightAvailability.FlightClass};{b.Passenger.Id}");
-
-            return stringBuilder.ToString();
-        }
+      
     }
 }
