+++ AirportTicketBookingSystem/FlightsInventory.cs
@@ -17,114 +17,6 @@ namespace AirportTicketBookingSystem
         public List<Country> Countries { get; } = new List<Country>();
         public List<Airport> Airports { get; } = new List<Airport>();
 
-
-
-        // ******************************************
-        // Load Data
-        // ******************************************
-
-        public void LoadFlightsData()
-        {
-            //Flights.Clear();
-            LoadCountries();
-            LoadAirports();
-            LoadFlights();
-            LoadFlightAvailability();
-        }
-
-        public void LoadCountries()
-        {
-            List<string> countryNames = ["Argentina", "Colombia", "Chile"];
-
-            foreach (string countryName in countryNames)
-            {
-                Countries.Add(new Country(countryName));
-            }
-        }
-
-        public void LoadAirports()
-        {
-            List<string> ArgentinaAirportNames = ["Aeroparque", "Ezeiza", "Ing. Ambrosio"];
-            List<string> ColombiaAirportNames = ["El Dorado", "Puente Aereo", "Olaya Herrera"];
-            List<string> ChileAirportNames = ["Santiago", "Araucania", "Arturo Merino Benitez"];
-
-            foreach (Country country in Countries)
-            {
-                if (country.Name.Equals("Argentina"))
-                {
-                    foreach (string airportName in ArgentinaAirportNames)
-                    {
-                        Airports.Add(new Airport(airportName, country));
-                    }
-
-                }
-                else if (country.Name.Equals("Colombia"))
-                {
-                    foreach (string airportName in ColombiaAirportNames)
-                    {
-                        Airports.Add(new Airport(airportName, country));
-                    }
-                }
-                else if (country.Name.Equals("Chile"))
-                {
-                    foreach (string airportName in ChileAirportNames)
-                    {
-                        Airports.Add(new Airport(airportName, country));
-                    }
-                }
-            }
-        }
-
-        public void LoadFlights()
-        {
-            List<FlightAvailability> flightAvailabilities = LoadFlightAvailability();
-            Flight f1 = new(1, new DateTime(2024, 5, 6, 10, 30, 0), Airports[0], Airports[2], flightAvailabilities);
-            Flight f2 = new(2, new DateTime(2024, 5, 7, 10, 30, 0), Airports[1], Airports[3], flightAvailabilities);
-            Flight f3 = new(3, new DateTime(2024, 5, 8, 10, 30, 0), Airports[2], Airports[4], flightAvailabilities);
-            Flight f4 = new(4, new DateTime(2024, 5, 9, 10, 30, 0), Airports[3], Airports[5], flightAvailabilities);
-            Flight f5 = new(5, new DateTime(2024, 5, 6, 22, 30, 0), Airports[4], Airports[6], flightAvailabilities);
-            Flight f6 = new(6, new DateTime(2024, 5, 7, 15, 30, 0), Airports[5], Airports[7], flightAvailabilities);
-            Flight f7 = new(7, new DateTime(2024, 5, 8, 20, 30, 0), Airports[6], Airports[8], flightAvailabilities);
-            Flight f8 = new(8, new DateTime(2024, 5, 9, 18, 30, 0), Airports[7], Airports[0], flightAvailabilities);
-            Flight f9 = new(9, new DateTime(2024, 5, 6, 10, 0, 0), Airports[8], Airports[1], flightAvailabilities);
-            Flight f10 = new(10, new DateTime(2024, 5, 7, 7, 30, 0), Airports[0], Airports[2], flightAvailabilities);
-
-            List<FlightAvailability> flightAvailabilities1 = LoadExtraFlightAvailability();
-            Flight f11 = new(11, new DateTime(2024, 5, 7, 7, 30, 0), Airports[0], Airports[8], flightAvailabilities1);
-
-            Flights.AddRange([f1, f2, f3, f4, f5, f6, f7, f8, f9, f10, f11]);
-
-        }
-
-        public List<FlightAvailability> LoadFlightAvailability()
-        {
-            List<int> placesByClass = [20, 10, 5];
-            List<double> pricesByClass = [1000, 1500, 2000];
-            List<FlightAvailability> flightsByClass = new List<FlightAvailability>();
-            foreach (FlightClass fc in Enum.GetValues(typeof(FlightClass)))
-            {
-                int places = placesByClass[(int)fc];
-                double price = pricesByClass[(int)fc];
-                flightsByClass.Add(new FlightAvailability(fc, price, places, places));
-            }
-            return flightsByClass;
-        }
-
-        public List<FlightAvailability> LoadExtraFlightAvailability()
-        {
-            List<int> placesByClass = [30, 15, 5];
-            List<double> pricesByClass = [3000, 3500, 4000];
-            List<FlightAvailability> flightsByClass = new List<FlightAvailability>();
-            foreach (FlightClass fc in Enum.GetValues(typeof(FlightClass)))
-            {
-                int places = placesByClass[(int)fc];
-                double price = pricesByClass[(int)fc];
-                flightsByClass.Add(new FlightAvailability(fc, price, places, places));
-            }
-            return flightsByClass;
-        }
-
-
         // ******************************************
         // Searches
         // ******************************************
